+++ src/jedison.js
@@ -68,7 +68,6 @@ class Jedison extends EventEmitter {
       translations: {},
       settings: {},
       useConstraintAttributes: true,
-      useStringConstraintAttributes: false,
       parseMarkdown: false,
       purifyHtml: true,
       purifyData: true,